requirements.txt
asgiref==3.2.7
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
gunicorn==20.0.4
pymongo==3.10.1
PyMySQL==0.9.3
pytz==2019.3
shortuuid==1.0.1
sqlparse==0.3.1
